% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_functions.R
\name{annotate_circle}
\alias{annotate_circle}
\title{annotate concentric circles}
\usage{
annotate_circle(cent, col = "#00BFFF", format = "", pos = "top",
  text_size = 3)
}
\arguments{
\item{cent}{centrality scores used for layout}

\item{col}{color of text}

\item{format}{either empty string or 'scientific'}

\item{pos}{position of text ('top' or 'bottom')}

\item{text_size}{font size for annotations}
}
\value{
concentric circles around origin
}
\description{
annotate concentric circles
}
