% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oaqc.R
\name{annotate_result}
\alias{annotate_result}
\title{Annotates the igraph object with orbit labels.}
\usage{
annotate_result(graph, orbits, non_ind_freq)
}
\arguments{
\item{graph}{Unmodified input graph.}

\item{orbits}{List with n_orbits, e_orbits matrices.}

\item{non_ind_freq}{A flag indicating whether non-induced frequencies have to be written or not.}
}
\value{
\code{orbits} if the input is not an igraph, the annotated igraph
instead.
}
\description{
Annotates the igraph object with orbit labels.
}
