% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oaqc.R
\name{as.edge_list}
\alias{as.edge_list}
\title{This files is adapted from the archived R package oaqc
Coerce graph input.}
\usage{
as.edge_list(graph)
}
\arguments{
\item{graph}{A matrix, data.frame or graph object.}
}
\value{
Edge list matrix.
}
\description{
This files is adapted from the archived R package oaqc
Coerce graph input.
}
