% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_functions.R
\name{draw_circle}
\alias{draw_circle}
\title{Draw concentric circles}
\usage{
draw_circle(col = "#00BFFF", use = "focus", max.circle)
}
\arguments{
\item{col}{color of circles}

\item{use}{one of 'focus' or 'cent'}

\item{max.circle}{if use='focus' specifies the number of circles to draw}
}
\value{
concentric circles around origin
}
\description{
Draw concentric circles
}
