% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manipulate_graph.R
\name{graph_manipulate}
\alias{graph_manipulate}
\alias{reorder_edges}
\title{Manipulate graph}
\usage{
reorder_edges(g, attr, desc = T)
}
\arguments{
\item{g}{igraph object}

\item{attr}{attribute to sort edges}

\item{desc}{logical. sort in descending (default) or ascending order}
}
\value{
manipulated graphs
}
\description{
functions to manipulate graphs, such as changing order of edges
}
\author{
David Schoch
}
