% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backbone_function.R
\name{layout_as_backbone}
\alias{layout_as_backbone}
\title{Backbone graph layout}
\usage{
layout_as_backbone(g, keep = 0.2, backbone = T)
}
\arguments{
\item{g}{igraph object}

\item{keep}{fraction of edges to keep in backbone calculation}

\item{backbone}{logical. Return edge ids of the backbone}
}
\value{
coordinates to be used layouting a graph
}
\description{
Backbone graph layout
}
