% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layout_stress.R, R/layouts.R
\name{layout_centrality_group}
\alias{layout_centrality_group}
\alias{layout_with_centrality_group}
\alias{layout_igraph_centrality_group}
\title{radial centrality group layout}
\usage{
layout_with_centrality_group(g, cent, group, shrink = 10, ...)

layout_igraph_centrality_group(g, cent, group, shrink = 10, circular, ...)
}
\arguments{
\item{g}{igraph object}

\item{cent}{centrality scores}

\item{group}{vector indicating grouping of nodes}

\item{shrink}{shrink the reserved angle range for a group to increase the gaps between groups}

\item{...}{additional arguments to layout_with_centrality
The layout_igraph_* function should not be used directly. It is only used as an argument for plotting with 'igraph'.
'ggraph' natively supports the layout.}

\item{circular}{not used}
}
\value{
matrix of xy coordinates
}
\description{
arranges nodes in concentric circles according to a centrality index and keeping groups within a angle range
}
\examples{
library(igraph)
}
\seealso{
\link{layout_centrality}
}
