% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manipulate_layout.R
\name{layout_manipulate}
\alias{layout_manipulate}
\alias{layout_rotate}
\alias{layout_mirror}
\title{Manipulate layout}
\usage{
layout_rotate(xy, angle)

layout_mirror(xy, axis = "vertical")
}
\arguments{
\item{xy}{graph layout}

\item{angle}{angle for rotation}

\item{axis}{mirror horizontal or vertical}
}
\value{
manipulated layout
}
\description{
functions to manipulate an existing layout
}
\author{
David Schoch
}
