% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stress_function.R
\name{layout_with_centrality}
\alias{layout_with_centrality}
\title{Centrality layout}
\usage{
layout_with_centrality(g, cent, iter = 500, tol = 1e-04,
  tseq = seq(0, 1, 0.2))
}
\arguments{
\item{g}{igraph object}

\item{cent}{centrality scores}

\item{iter}{number of iterations}

\item{tol}{stoping criterion}

\item{tseq}{transition steps}
}
\value{
coordinates to be used layouting a graph
}
\description{
Centrality layout
}
