% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stress_function.R
\name{layout_with_focus}
\alias{layout_with_focus}
\title{Focal layout}
\usage{
layout_with_focus(g, v, iter = 500, tol = 1e-04)
}
\arguments{
\item{g}{igraph object}

\item{v}{focal node to be placed in the center}

\item{iter}{number of iterations}

\item{tol}{stoping criterion}
}
\value{
coordinates to be used layouting a graph
}
\description{
Focal layout
}
