% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-examples.R
\docType{data}
\name{metro_berlin}
\alias{metro_berlin}
\title{Subway network of Berlin}
\format{
igraph object
}
\usage{
metro_berlin
}
\description{
A dataset containing the subway network of Berlin
}
\references{
Kujala, Rainer, et al. "A collection of public transport network data sets for 25 cities." Scientific data 5 (2018): 180089.
}
\keyword{datasets}
