% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-examples.R
\docType{data}
\name{multilvl_ex}
\alias{multilvl_ex}
\title{Multilevel example Network}
\format{
igraph object
}
\usage{
multilvl_ex
}
\description{
Multilevel network, where both levels have different structural features
}
\keyword{datasets}
