% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/large_graphs.R, R/layouts.R
\name{pivotMDS}
\alias{pivotMDS}
\alias{layout_with_pmds}
\alias{layout_igraph_pmds}
\alias{layout_tbl_graph_pmds}
\title{pivot MDS graph layout}
\usage{
layout_with_pmds(g, pivots, weights = NA)

layout_igraph_pmds(g, pivots, weights = NA, circular)

layout_tbl_graph_pmds(g, pivots, weights = NA, circular)
}
\arguments{
\item{g}{igraph object}

\item{pivots}{number of pivots}

\item{weights}{Possibly a numeric vector with edge weights. If this is NULL and the graph has a weight edge attribute, then the attribute is used. If this is NA then no weights are used (even if the graph has a weight attribute). By default, weights are ignored. See details for more.}

\item{circular}{not used}

\item{circular}{not used}
}
\value{
coordinates to be used layouting a graph
}
\description{
similar to MDS layout but uses only a small set of pivots for MDS. Considerably faster than MDS and thus applicable for larger graphs.
}
\details{
the layout_igraph_* function should not be used directly. It is only used as an argument for 'ggraph'.
Be careful when using weights. In most cases, the inverse of the edge weights should be used to ensure that the endpoints of an edges with higher weights are closer together (weights=1/E(g)$weight)
}
\examples{
\dontrun{
library(igraph)
library(ggraph)
g <- sample_gnp(1000,0.01)
ggraph(g,layout="pMDS",pivots=250)+
  geom_edge_link(n=2,edge_colour="grey66")+
  geom_node_point(shape=21,fill="grey25",size=5)+
  theme_graph()
}
}
\references{
Brandes, U. and Pich, C. (2006). Eigensolver Methods for Progressive Multidimensional Scaling of Large Data. In \emph{International Symposium on Graph Drawing} (pp. 42-53). Springer
}
\author{
David Schoch
}
