% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layouts.R
\name{qgraph}
\alias{qgraph}
\title{Quick plot network}
\usage{
qgraph(g)
}
\arguments{
\item{g}{igraph object}
}
\description{
qgraph is a shortcut designed to obtain a quick view of a network using a stress based layout.
}
