% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/large_graphs.R, R/layouts.R
\name{sparseStress}
\alias{sparseStress}
\alias{layout_with_sparseStress}
\alias{layout_igraph_sparseStress}
\title{sparse stress graph layout}
\usage{
layout_with_sparseStress(g, pivots)

layout_igraph_sparseStress(g, pivots, circular)
}
\arguments{
\item{g}{igraph object}

\item{pivots}{number of pivots}

\item{circular}{not used}
}
\value{
coordinates to be used layouting a graph
}
\description{
stress majorization for larger graphs
}
\details{
the layout_igraph_* function should not be used directly. It is only used as an argument for 'ggraph'. Edge weights are not suported yet.
}
\examples{
\dontrun{
library(igraph)
library(ggraph)
g <- sample_gnp(1000,0.005)
ggraph(g,layout="sparseStress",pivots=250)+
  geom_edge_link(n=2,edge_colour="grey66")+
  geom_node_point(shape=21,fill="grey25",size=5)+
  theme_graph()
}
}
\references{
Ortmann, M. and Klimenta, M. and Brandes, U. (2016).A Sparse Stress Model. https://arxiv.org/pdf/1608.08909.pdf
}
