% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stress_function.R
\name{stress_majorization}
\alias{stress_majorization}
\title{Stress majorization graph layout}
\usage{
stress_majorization(g, iter = 500, tol = 1e-04, mds = TRUE, bbox = 10)
}
\arguments{
\item{g}{igraph object}

\item{iter}{number of iterations}

\item{tol}{stoping criterion}

\item{mds}{should an MDS layout be used as initial layout (default: TRUE)}

\item{bbox}{constrain dimension of output}
}
\value{
coordinates to be used layouting a graph
}
\description{
Stress majorization graph layout
}
