% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backbone_function.R
\name{backbone_layout}
\alias{backbone_layout}
\title{Backbone layout graph layout}
\usage{
backbone_layout(g, keep = 0.2, backbone = T)
}
\arguments{
\item{g}{igraph object}

\item{keep}{fraction of edges to keep}

\item{backbone}{logical. Return edge ids of the backbone}
}
\value{
coordinates to be used layouting a graph
}
\description{
Backbone layout graph layout
}
