% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynamic_graphs.R
\name{dynamic_layout}
\alias{dynamic_layout}
\alias{layout_as_dynamic}
\title{dynamic graph layout}
\usage{
layout_as_dynamic(gList, alpha = 0.5, iter = 500, tol = 1e-04)
}
\arguments{
\item{gList}{list of igraph objects. Each network must contain the same set of nodes.}

\item{alpha}{weighting of reference layout. See details.}

\item{iter}{number of iterations during stress optimization}

\item{tol}{stopping criterion for stress optimization}
}
\value{
list of coordinates for each graph
}
\description{
Create layouts for longitudinal networks.
}
\details{
The reference layout is calculated based on the union of all graphs. The parameter alpha controls the influence of the reference layout.
For alpha=1, only the reference layout is used and all graphs have the same layout. For alpha=0, the stress layout of each individual graph is used. Values inbetween interpolate between the two layouts.
}
\examples{
library(igraph)
g1 <- sample_gnp(20,0.2)
g2 <- sample_gnp(20,0.2)
g3 <- sample_gnp(20,0.2)

xy <- layout_as_dynamic(list(g1,g2,g3))

# layout for first network
xy[[1]]
}
\references{
Brandes, U. and IndleKofer, N. and Mader, M. (2012). Visualization methods for longitudinal social networks and stochastic actor-oriented modeling. \emph{Social Networks} 34 (3) 291-308
}
