% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stress_function.R
\name{stress_majorization}
\alias{stress_majorization}
\title{Stress majorization graph layout}
\usage{
stress_majorization(g, iter = 500, tol = 1e-04)
}
\arguments{
\item{g}{igraph object}

\item{iter}{number of iterations}

\item{tol}{stoping criterion}

\item{dim}{dimension of output coordinates. default is 2}
}
\value{
coordinates to be used layouting a graph
}
\description{
Stress majorization graph layout
}
